Evidencia3/app/requirements.txt
pandas==2.1.1
matplotlib==3.8.0
mysql-connector-python==8.0.33

